:recycle: refactor: [CodeRabbit 피드백 반영]일반 예외 메시지 노출 제거 및 패키지명 오타 수정
