+++ src/main/java/com/example/tracklybe/global/exception/HabitNotFoundException.java
@@ -1,6 +1,6 @@
 package com.example.tracklybe.global.exception;
 
-import com.example.tracklybe.global.exception.enummeration.ErrorCode;
+import com.example.tracklybe.global.exception.enumeration.ErrorCode;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
